+++ rust/numaflow-sideinput/src/error.rs
@@ -4,6 +4,9 @@ pub type Result<T> = std::result::Result<T, Error>;
 
 #[derive(Error, Debug, Clone)]
 pub enum Error {
+    #[error("Schedule Error - {0}")]
+    Schedule(String),
+
     #[error("SideInput Error - {0}")]
     SideInput(String),
